requirements.txt
APScheduler==3.9.1
betfairlightweight==2.16.4
dropbox==11.30.0
joblib==1.1.0
numpy==1.22.3
pandas==1.4.2
python_dateutil==2.8.2
requests==2.27.1
tenacity==8.0.1
mysql-connector==2.2.9
lightgbm==3.3.2

